+++ src/Event/AbstractListener.php
@@ -1,0 +1,21 @@
+<?php
+/**
+ * Created by PhpStorm.
+ * User: fsilva
+ * Date: 21/08/2018
+ * Time: 00:13
+ */
+
+namespace Slick\CQRSTools\Event;
+
+use League\Event\AbstractListener as LeagueListener;
+
+/**
+ * Abstract Listener
+ *
+ * @package Slick\CQRSTools\Event
+ */
+abstract class AbstractListener extends LeagueListener
+{
+
+}